refactor(print): reorganize internals and align XR comment style
