+++ src/core/print/print_api.hpp
@@ -13,8 +13,7 @@
 namespace LibXR::Print
 {
 /**
- * @brief Public return-value contract for the print convenience surface.
- * @brief print 便捷接口的公开返回值约定。
+ * @brief print 便捷接口的公开返回值约定 / Public return-value contract for the print convenience surface
  *
  * Sink-writing helpers such as Write(), FormatTo(), and PrintfTo() return
  * ErrorCode only and report success or failure without exposing a length.
@@ -36,9 +35,14 @@ namespace LibXR::Print
  */
 
 /**
- * @brief Writes a compiled format into a sink.
- * @brief 将编译后的格式写入输出端。
- * @return Returns the sink write status only. / 仅返回 sink 写入状态。
+ * @brief 将编译后的格式写入输出端 / Write one compiled format into a sink
+ * @tparam Sink 输出端类型，需满足 `OutputSink` / Sink type satisfying `OutputSink`
+ * @tparam Format 编译格式类型，需满足 `CompiledFormat` / Compiled format type satisfying `CompiledFormat`
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param sink 输出端 / Destination sink
+ * @param format 编译后的格式对象 / Compiled format object
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 仅返回 sink 写入状态 / Returns only the sink write status
  */
 template <typename Sink, typename Format, typename... Args>
 requires OutputSink<Sink> && CompiledFormat<std::remove_cvref_t<Format>>
@@ -50,9 +54,14 @@ requires OutputSink<Sink> && CompiledFormat<std::remove_cvref_t<Format>>
 }
 
 /**
- * @brief Public convenience wrapper that writes one compiled format into one sink.
- * @brief 面向用户的便捷包装：将一份编译格式写入一个输出端。
- * @return Returns the sink write status only. / 仅返回 sink 写入状态。
+ * @brief 面向用户的便捷包装：将一份编译格式写入一个输出端 / Public convenience wrapper that writes one compiled format into one sink
+ * @tparam Sink 输出端类型，需满足 `OutputSink` / Sink type satisfying `OutputSink`
+ * @tparam Format 编译格式类型 / Compiled format type
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param sink 输出端 / Destination sink
+ * @param format 编译后的格式对象 / Compiled format object
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 仅返回 sink 写入状态 / Returns only the sink write status
  */
 template <OutputSink Sink, typename Format, typename... Args>
 [[nodiscard]] inline ErrorCode FormatTo(Sink& sink, const Format& format, Args&&... args)
@@ -61,9 +70,13 @@ template <OutputSink Sink, typename Format, typename... Args>
 }
 
 /**
- * @brief Writes one brace-style literal directly into one sink.
- * @brief 将一条 brace 风格字面量直接写入一个输出端。
- * @return Returns the sink write status only. / 仅返回 sink 写入状态。
+ * @brief 将一条 brace 风格字面量直接写入一个输出端 / Write one brace-style literal directly into one sink
+ * @tparam Source brace 风格格式串字面量 / Brace-style format literal
+ * @tparam Sink 输出端类型，需满足 `OutputSink` / Sink type satisfying `OutputSink`
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param sink 输出端 / Destination sink
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 仅返回 sink 写入状态 / Returns only the sink write status
  */
 template <Text Source, OutputSink Sink, typename... Args>
 [[nodiscard]] inline ErrorCode FormatTo(Sink& sink, Args&&... args)
@@ -73,9 +86,13 @@ template <Text Source, OutputSink Sink, typename... Args>
 }
 
 /**
- * @brief Writes one printf-style literal directly into one sink.
- * @brief 将一条 printf 风格字面量直接写入一个输出端。
- * @return Returns the sink write status only. / 仅返回 sink 写入状态。
+ * @brief 将一条 printf 风格字面量直接写入一个输出端 / Write one printf-style literal directly into one sink
+ * @tparam Source printf 风格格式串字面量 / Printf-style format literal
+ * @tparam Sink 输出端类型，需满足 `OutputSink` / Sink type satisfying `OutputSink`
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param sink 输出端 / Destination sink
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 仅返回 sink 写入状态 / Returns only the sink write status
  */
 template <Text Source, OutputSink Sink, typename... Args>
 [[nodiscard]] inline ErrorCode PrintfTo(Sink& sink, Args&&... args)
@@ -85,9 +102,11 @@ template <Text Source, OutputSink Sink, typename... Args>
 }
 
 /**
- * @brief Formats one compiled format into one bounded char buffer and keeps the
- *        result NUL-terminated when capacity is nonzero.
- * @brief 将一份编译格式写入一个有界 char 缓冲区；当容量非零时，结果始终带结尾零字节。
+ * @brief 将一份编译格式写入一个有界 char 缓冲区；当容量非零时结果始终带结尾零字节 / Format one compiled format into one bounded char buffer and keep the result NUL-terminated when capacity is nonzero
+ * @param buffer 目标缓冲区；当 `capacity == 0` 时允许为空 / Destination buffer; may be null when `capacity == 0`
+ * @param capacity 目标缓冲区总容量（字节） / Total destination capacity in bytes
+ * @param format 编译后的格式对象 / Compiled format object
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
  *
  * The returned size follows snprintf semantics: it is the full character count
  * that would have been produced without truncation, excluding the trailing zero
@@ -100,6 +119,10 @@ template <Text Source, OutputSink Sink, typename... Args>
  * When capacity is nonzero, the destination buffer is always kept
  * NUL-terminated.
  * 当 capacity 非零时，目标缓冲区始终保持 NUL 结尾。
+ *
+ * @note `buffer` 真正最多保留 `capacity - 1` 个可见字符；最后一个字节保留给结尾零 / The buffer retains at most `capacity - 1` visible characters; the last byte is reserved for the trailing zero
+ * @note 当 `capacity == 0` 时不会写入结尾零，返回值仍然表示完整未截断长度 / When `capacity == 0`, no trailing zero is written and the return value still reports the full untruncated size
+ * @return 成功时返回完整格式化长度（不含结尾零字节）；运行期错误或长度超出 `int` 可表示范围时返回 `-1` / Returns the full formatted size excluding the trailing zero byte on success, or `-1` on runtime error or `int` overflow
  */
 [[nodiscard]] inline int FormatIntoBuffer(char* buffer, size_t capacity, const auto& format,
                                           auto&&... args)
@@ -159,11 +182,13 @@ template <Text Source, OutputSink Sink, typename... Args>
 }
 
 /**
- * @brief Formats one brace-style literal directly into one bounded char buffer.
- * @brief 将一条 brace 风格字面量直接写入一个有界 char 缓冲区。
- * @return Returns the full formatted size excluding the trailing zero byte, or
- *         -1 on runtime error (including size overflow).
- *         返回完整格式化长度，不含结尾零字节；运行期错误（包括尺寸溢出）返回 -1。
+ * @brief 将一条 brace 风格字面量直接写入一个有界 char 缓冲区 / Format one brace-style literal directly into one bounded char buffer
+ * @tparam Source brace 风格格式串字面量 / Brace-style format literal
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param buffer 目标缓冲区；当 `capacity == 0` 时允许为空 / Destination buffer; may be null when `capacity == 0`
+ * @param capacity 目标缓冲区总容量（字节） / Total destination capacity in bytes
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 返回完整格式化长度，不含结尾零字节；运行期错误（包括尺寸溢出）返回 `-1` / Returns the full formatted size excluding the trailing zero byte, or `-1` on runtime error including size overflow
  */
 template <Text Source, typename... Args>
 [[nodiscard]] inline int FormatIntoBuffer(char* buffer, size_t capacity, Args&&... args)
@@ -173,11 +198,13 @@ template <Text Source, typename... Args>
 }
 
 /**
- * @brief Formats one printf-style literal directly into one bounded char buffer.
- * @brief 将一条 printf 风格字面量直接写入一个有界 char 缓冲区。
- * @return Returns the full formatted size excluding the trailing zero byte, or
- *         -1 on runtime error (including size overflow).
- *         返回完整格式化长度，不含结尾零字节；运行期错误（包括尺寸溢出）返回 -1。
+ * @brief 将一条 printf 风格字面量直接写入一个有界 char 缓冲区 / Format one printf-style literal directly into one bounded char buffer
+ * @tparam Source printf 风格格式串字面量 / Printf-style format literal
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param buffer 目标缓冲区；当 `capacity == 0` 时允许为空 / Destination buffer; may be null when `capacity == 0`
+ * @param capacity 目标缓冲区总容量（字节） / Total destination capacity in bytes
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 返回完整格式化长度，不含结尾零字节；运行期错误（包括尺寸溢出）返回 `-1` / Returns the full formatted size excluding the trailing zero byte, or `-1` on runtime error including size overflow
  */
 template <Text Source, typename... Args>
 [[nodiscard]] inline int PrintfIntoBuffer(char* buffer, size_t capacity, Args&&... args)
@@ -187,14 +214,20 @@ template <Text Source, typename... Args>
 }
 
 /**
- * @brief snprintf-style wrapper built on top of the compiled-format path.
- * @brief 基于编译格式路径实现的 snprintf 风格包装。
+ * @brief 基于编译格式路径实现的 snprintf 风格包装 / snprintf-style wrapper built on top of the compiled-format path
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param buffer 目标缓冲区；当 `capacity == 0` 时允许为空 / Destination buffer; may be null when `capacity == 0`
+ * @param capacity 目标缓冲区总容量（字节） / Total destination capacity in bytes
+ * @param format 编译后的格式对象 / Compiled format object
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
  *
  * Returns the full formatted size excluding the trailing zero byte, or -1 on
  * runtime error (including size overflow).
  * 返回完整格式化长度（不含结尾零字节）；运行期错误（包括尺寸溢出）返回 -1。
  * Truncation is not an error and still returns the full size.
  * 截断不算错误，返回值仍然是完整长度。
+ *
+ * @return 成功时返回完整格式化长度；运行期错误或长度超出 `int` 可表示范围时返回 `-1` / Returns the full formatted size on success, or `-1` on runtime error or `int` overflow
  */
 [[nodiscard]] inline int SNPrintf(char* buffer, size_t capacity, const auto& format,
                                   auto&&... args)
@@ -203,11 +236,13 @@ template <Text Source, typename... Args>
 }
 
 /**
- * @brief snprintf-style wrapper for one printf-style literal.
- * @brief 面向一条 printf 风格字面量的 snprintf 风格包装。
- * @return Returns the full formatted size excluding the trailing zero byte, or
- *         -1 on runtime error (including size overflow).
- *         返回完整格式化长度，不含结尾零字节；运行期错误（包括尺寸溢出）返回 -1。
+ * @brief 面向一条 printf 风格字面量的 snprintf 风格包装 / snprintf-style wrapper for one printf-style literal
+ * @tparam Source printf 风格格式串字面量 / Printf-style format literal
+ * @tparam Args 调用点实参类型列表 / Call-site argument types
+ * @param buffer 目标缓冲区；当 `capacity == 0` 时允许为空 / Destination buffer; may be null when `capacity == 0`
+ * @param capacity 目标缓冲区总容量（字节） / Total destination capacity in bytes
+ * @param args 参与本次格式化的调用点实参 / Call-site arguments used for this formatting pass
+ * @return 返回完整格式化长度，不含结尾零字节；运行期错误（包括尺寸溢出）返回 `-1` / Returns the full formatted size excluding the trailing zero byte, or `-1` on runtime error including size overflow
  */
 template <Text Source, typename... Args>
 [[nodiscard]] inline int SNPrintf(char* buffer, size_t capacity, Args&&... args)
